+++ .github/workflows/tests.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Test
         run: npm run test
       - name: Test browsers
-        run: npx jackrabbit_webdriver .github/workflows/browsers.json ./tests/dist/mod.js
+        run: npx jackrabbit_webdriver .github/workflows/browser.json ./tests/dist/mod.js
   # build_and_test_macos:
   #   runs-on: macos-latest
   #   steps:
